+++ NEWS.md
@@ -3,6 +3,7 @@
 - rewrite getchain to speed up
 - add shiny application pmdnet to perform PMD network analysis
 - add support for formula in getchain to find compounds ions
+- remove frequency cutoff in getsda and use PMD network clusters analysis to determine the cutoff
 
 # pmd 0.1.5